refactor: 사용하지 않는 exception 제거 및 패키지 이동
