+++ src/main/java/global/exception/AbstractExceptionHandler.java
@@ -1,4 +1,4 @@
-package global;
+package global.exception;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
